refactor(logging): 重构日志输出逻辑并更新Docker配置- 重构访问日志和通用日志输出逻辑。现在日志输出到文件和stdout的条件更加清晰。 - 更新Dockerfile，修正版本标签格式并解决文件添加路径问题。 - 在Makefile中新增init-db目标，方便初始化数据库。 - 清理setup.go文件，删除不必要的日志读取操作。
次改动提升了代码可维护性，并确保日志策略符合预期行为。此外，构建和数据库初始化流程得到改进。

+++ logger/access.go
@@ -70,14 +70,20 @@ func NewRequestLogger(dot func(...interface{})) (h iris.Handler, close func() er
 		Path:              true,
 		MessageHeaderKeys: []string{"User-Agent", "traceId"},
 	}
-	logFile = newLogFile()
-	close = func() error {
-		err := logFile.Close()
-		if deleteFileOnExit {
-			err = os.Remove(logFile.Name())
+
+	if isLogOutFile() {
+		logFile = newLogFile()
+		close = func() error {
+			err := logFile.Close()
+			if deleteFileOnExit {
+				err = os.Remove(logFile.Name())
+			}
+			return err
 		}
-		return err
+	} else {
+		logFile = os.Stdout
 	}
+
 	c.LogFuncCtx = func(ctx iris.Context, latency time.Duration) {
 		traceId := ctx.Request().Header.Get("traceId")
 		if traceId == "" {
